requirements.txt
fastapi
uvicorn
slowapi
python-jose[cryptography]
passlib
python-dotenv
google-generativeai
duckduckgo-search
pydantic
pydantic-settings
